requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
langchain==0.1.5
langchain-community==0.0.20
pdfplumber==0.10.3
pdf2image==1.16.3
pytesseract==0.3.10
python-docx==1.1.0
Pillow==10.2.0
uuid==1.30
transformers==4.38.2
torch==2.2.1
sentence-transformers==2.5.1
ollama==0.1.7
